+++ src/test/java/org/heigit/ohsome/ohsomeapi/controller/PostControllerTest.java
@@ -45,7 +45,7 @@ public class PostControllerTest {
   /** Stops this application context. */
   @AfterClass
   public static void applicationMainShutdown() {
-    if (null != Application.getApplicationContext()) {
+    if (Application.getApplicationContext()  != null) {
       SpringApplication.exit(Application.getApplicationContext(), () -> 0);
     }
   }